Implementación de sistema de Suscripción y Productos Digitales

+++ src/Controllers/ProductController.php
@@ -12,6 +12,7 @@ use Carbon\Carbon;
 use Nowyouwerkn\WeCommerce\Models\Product;
 use Nowyouwerkn\WeCommerce\Models\Category;
 use Nowyouwerkn\WeCommerce\Models\ProductImage;
+use Nowyouwerkn\WeCommerce\Models\ProductCharacteristic;
 use Nowyouwerkn\WeCommerce\Models\ProductVariant;
 use Nowyouwerkn\WeCommerce\Models\ProductRelationship;
 
@@ -51,7 +52,23 @@ class ProductController extends Controller
         ->with('categories', $categories);
     }
 
-        public function promotions()
+    public function createDigital()
+    {
+        $categories = Category::where('parent_id', NULL)->orWhere('parent_id', '0')->get();
+
+        return view('wecommerce::back.products.create_digital')
+        ->with('categories', $categories);
+    }
+
+    public function createSubscription()
+    {
+        $categories = Category::where('parent_id', NULL)->orWhere('parent_id', '0')->get();
+
+        return view('wecommerce::back.products.create_subscription')
+        ->with('categories', $categories);
+    }
+
+    public function promotions()
     {
         $products = Product::where('has_discount', true)->paginate(15);
 
@@ -61,13 +78,23 @@ class ProductController extends Controller
     public function store(Request $request)
     {
         //Validar
-        $this -> validate($request, array(
-            'name' => 'unique:products|required|max:255',
-            'description' => 'required|min:30',
-            'price' => 'required',
-            'model_image' => 'sometimes|min:10|max:2100|image',
-            'sku' => 'required',
-        ));
+        if($request->type == 'subscription'){
+            $this -> validate($request, array(
+                'name' => 'unique:products|required|max:255',
+                'description' => 'required|min:30',
+                'price' => 'required',
+                'model_image' => 'sometimes|min:10|max:2100|image',
+            ));
+        }else{
+            $this -> validate($request, array(
+                'name' => 'unique:products|required|max:255',
+                'description' => 'required|min:30',
+                'price' => 'required',
+                'model_image' => 'sometimes|min:10|max:2100|image',
+                'sku' => 'required',
+            ));
+        }
+        
 
         /* Crear categoría si usuario activó opción */
         if ($request->category_name != NULL) {
@@ -82,29 +109,22 @@ class ProductController extends Controller
         // Guardar datos en la base de datos
         $product = new Product;
 
+        /*  Características Generales */
         $product->name = $request->name;
         $product->slug = Str::slug($request->name);
         $product->description = $request->description;
-        $product->materials = $request->materials;
+        $product->materials = $request->materials ?? NULL;
+        $product->terms_conditions = $request->terms_conditions ?? NULL;
+        
         $product->color = $request->color;
         $product->hex_color = $request->hex_color;
         $product->pattern = $request->pattern;
-
         $product->in_index = $request->in_index;
         $product->is_favorite = $request->is_favorite;
 
         $product->price = $request->price;
-          if ($request->discount_price != NULL) {
-             $product->discount_price =  $request->discount_price;
-        }else{
-             $product->discount_price = $request->discount_price;
-        }
-
-        if ($request->production_cost != NULL) {
-              $product->production_cost = $request->production_cost;
-        }else{
-             $product->production_cost = $request->production_cost;
-        }
+        $product->discount_price = $request->discount_price;
+        $product->production_cost = $request->production_cost;
         
         $product->has_discount = $request->has_discount;
         $product->discount_start = Carbon::now()->format('Y-m-d');
@@ -117,8 +137,8 @@ class ProductController extends Controller
         $product->stock = $request->stock;
 
         $product->has_variants = $request->has_variants;
-
         $product->size_chart_file = $request->size_chart_file;
+
         $product->height = $request->height;
         $product->width = $request->width;
         $product->lenght = $request->lenght;
@@ -139,11 +159,11 @@ class ProductController extends Controller
         $product->availability = $request->availability;
         $product->visibility = $request->visibility;
         $product->condition = $request->condition;
-        $product->product_type = $request->product_type;
+        
+        $product->type = $request->type;
+
         $product->fb_product_category = $request->fb_product_category;
         $product->google_product_category = $request->google_product_category;
-
-
         $product->available_date_start = $request->available_date_start;
 
         if ($request->hasFile('model_image')) {
@@ -156,18 +176,91 @@ class ProductController extends Controller
             $product->image = $filename;
         }
 
-        $product->save();
+        /* Características Productos Digitales */
+        if($request->type == 'digital'){
+            $product->download_link = $request->download_link;
 
+            /* Crear Slug del Nombre */
+            $nameslug = Str::slug($request->name);
+            $code_gen = substr(md5(uniqid(mt_rand(), true)) , 0, 5);
+
+            if ($request->hasFile('doc_file')) {
+                $archivo = $request->file('doc_file');
+                $filename = $nameslug . '.' . $archivo->getClientOriginalExtension();
+                
+                $location = public_path('files/products/');
+                $archivo->move($location, $filename);
+
+                $product->doc_file = $filename;
+            }
+
+            if ($request->hasFile('image_file')) {
+                $archivo = $request->file('image_file');
+                $filename = $nameslug . '.' . $archivo->getClientOriginalExtension();
+                
+                $location = public_path('files/products/');
+                $archivo->move($location, $filename);
+
+                $product->image_file = $filename;
+            }
+
+            $product->sku = $request->sku;
+            $product->barcode = $request->barcode;
+            $product->stock = '1';
+        }
+
+        /* Características Productos Suscripción */
+        if($request->type == 'subscription'){
+            $product->payment_frequency = $request->payment_frequency;
+            $product->payment_frequency_qty = $request->payment_frequency_qty;
+            $product->time_for_cancellation = $request->time_for_cancellation;
+
+            $code_gen = substr(md5(uniqid(mt_rand(), true)) , 0, 5);
+
+            $product->sku = (Str::slug($request->name) . '-' .$code_gen);
+            $product->barcode = (Str::slug($request->name) . '-' . $code_gen);
+            $product->stock = '1';
+            
+            // Guardar Características de producto configuradas
+            if(isset($request->char_title)){
+                $product->save();
+            
+                $characteristics = $request->char_title;
+
+                $index = 0;
+
+                foreach($characteristics as $char){
+                    $chars = new ProductCharacteristic;
+                    $chars->product_id = $product->id;
+                    $chars->title = $char;
+
+                    if(isset($request->char_subtitle)){
+                        $chars->subtitle = $request->char_subtitle[$index];
+                    }
+
+                    if(isset($request->char_icon)){
+                        $chars->icon = $request->char_icon[$index];
+                    }
+                    
+                    $chars->save();
+                    $index++;
+                }
+            }
+        }
+        
+        $product->save();
         $product->subCategory()->sync($request->subcategory);
 
         // Notificación
         $type = 'Producto';
         $by = Auth::user();
-        $data = 'creó el nuevo producto con nombre: ' . $product->name;
+        if($request->type == 'subscription'){
+            $data = 'creó un producto de suscripción con nombre: ' . $product->name;
+        }else{
+            $data = 'creó el nuevo producto con nombre: ' . $product->name;
+        }
         $model_action = "create";
         $model_id = $product->id;
-
-
 
         $this->notification->send($type, $by ,$data, $model_action, $model_id);
 
@@ -184,11 +277,8 @@ class ProductController extends Controller
         $product = Product::findOrFail($id);
 
         $categories = Category::where('parent_id', NULL)->orWhere('parent_id', '0')->get();
-
         $variant_stock = ProductVariant::where('product_id', $product->id)->get();
-
         $related_products = Product::where('id', '!=' , $id)->where('category_id', $product->category_id)->get();
-
         $total_qty = 0;
 
         foreach ($variant_stock as $v_stock) {
@@ -196,7 +286,6 @@ class ProductController extends Controller
         };
 
         $total_qty;
-
         $product_relationships = ProductRelationship::where('base_product_id', $id)->orWhere('product_id', $id)->get();
 
         if ($product_relationships->count() == NULL) {
@@ -206,7 +295,6 @@ class ProductController extends Controller
             $base_product = $product_relationships->take(1)->first();
             $all_relationships = ProductRelationship::where('base_product_id', $base_product->base_product_id)->get();
         }
-
 
         return view('wecommerce::back.products.show')
         ->with('product', $product)
@@ -356,21 +444,23 @@ class ProductController extends Controller
         // Guardar datos en la base de datos
         $product = Product::find($id);
 
+        /*  Características Generales */
         $product->name = $request->name;
         $product->slug = Str::slug($request->name);
         $product->description = $request->description;
-        $product->materials = $request->materials;
+        $product->materials = $request->materials ?? NULL;
+        $product->terms_conditions = $request->terms_conditions ?? NULL;
+        
         $product->color = $request->color;
         $product->hex_color = $request->hex_color;
         $product->pattern = $request->pattern;
-
         $product->in_index = $request->in_index;
         $product->is_favorite = $request->is_favorite;
 
         $product->price = $request->price;
-        $product->discount_price =  $request->discount_price;
+        $product->discount_price = $request->discount_price;
         $product->production_cost = $request->production_cost;
-
+        
         $product->has_discount = $request->has_discount;
         $product->discount_start = Carbon::now()->format('Y-m-d');
         $product->discount_end = $request->discount_end;
@@ -382,14 +472,18 @@ class ProductController extends Controller
         $product->stock = $request->stock;
 
         $product->has_variants = $request->has_variants;
-
         $product->size_chart_file = $request->size_chart_file;
+
         $product->height = $request->height;
         $product->width = $request->width;
         $product->lenght = $request->lenght;
         $product->weight = $request->weight;
 
-        $product->category_id = $request->category_id;
+        if ($request->category_name != NULL) {
+            $product->category_id= $category->id;
+        }else{
+            $product->category_id = $request->category_id;
+        }
 
         $product->status = $request->status;
         $product->search_tags = $request->search_tags;
@@ -401,11 +495,10 @@ class ProductController extends Controller
         $product->visibility = $request->visibility;
         $product->condition = $request->condition;
         
-        $product->product_type = $request->product_type;
+        //$product->type = $request->type;
 
         $product->fb_product_category = $request->fb_product_category;
         $product->google_product_category = $request->google_product_category;
-
         $product->available_date_start = $request->available_date_start;
 
         if ($request->hasFile('model_image')) {
@@ -418,6 +511,82 @@ class ProductController extends Controller
             $product->image = $filename;
         }
 
+        /* Características Productos Digitales */
+        if($product->type == 'digital'){
+            $product->download_link = $request->download_link;
+            //$product->doc_file = $request->doc_file;
+            //$product->image_file = $request->image_file;
+
+            /* Crear Slug del Nombre */
+            $nameslug = Str::slug($request->name);
+            $code_gen = substr(md5(uniqid(mt_rand(), true)) , 0, 5);
+
+            if ($request->hasFile('doc_file')) {
+                $archivo = $request->file('doc_file');
+                $filename = $nameslug . '.' . $archivo->getClientOriginalExtension();
+                
+                $location = public_path('files/products/');
+                $archivo->move($location, $filename);
+
+                $product->doc_file = $filename;
+            }
+
+            if ($request->hasFile('image_file')) {
+                $archivo = $request->file('image_file');
+                $filename = $nameslug . '.' . $archivo->getClientOriginalExtension();
+                
+                $location = public_path('files/products/');
+                $archivo->move($location, $filename);
+
+                $product->image_file = $filename;
+            }
+
+            $product->sku = $request->sku;
+            $product->barcode = $request->barcode;
+            $product->stock = '1';
+        }
+
+        /* Características Productos Suscripción */
+        if($product->type == 'subscription'){
+            $product->payment_frequency = $request->payment_frequency;
+            $product->payment_frequency_qty = $request->payment_frequency_qty;
+            $product->time_for_cancellation = $request->time_for_cancellation;
+
+            $code_gen = substr(md5(uniqid(mt_rand(), true)) , 0, 5);
+
+            $product->sku = (Str::slug($request->name) . '-' .$code_gen);
+            $product->barcode = (Str::slug($request->name) . '-' . $code_gen);
+            $product->stock = '1';
+            
+            // Guardar Características de producto configuradas
+            if(isset($request->char_title)){
+                $product->save();
+            
+                $characteristics = $request->char_title;
+
+                $index = 0;
+
+                foreach($characteristics as $char){
+                    $chars = new ProductCharacteristic;
+                    $chars->product_id = $product->id;
+                    $chars->title = $char;
+
+                    if(isset($request->char_subtitle)){
+                        $chars->subtitle = $request->char_subtitle[$index];
+                    }
+
+                    if(isset($request->char_icon)){
+                        $chars->icon = $request->char_icon[$index];
+                    }
+                    
+                    $chars->save();
+
+                    $index++;
+                }
+            }
+            
+        }
+        
         $product->save();
 
         if (isset($request->subcategory)) {
@@ -445,13 +614,17 @@ class ProductController extends Controller
         $product = Product::find($id);
         $vars = ProductVariant::where('product_id', $product->id)->get();
 
-        foreach($vars as $var){
-            $var->delete();
+        if($vars->count() != 0){
+            foreach($vars as $var){
+                $var->delete();
+            }
         }
-
+    
         $relations = ProductRelationship::where('product_id', $product->id)->orWhere('base_product_id', $product->id)->get();
-        foreach($relations as $rel){
-            $rel->delete();
+        if($relations->count() != 0){
+            foreach($relations as $rel){
+                $rel->delete();
+            }
         }
 
         // Notificación
@@ -461,10 +634,8 @@ class ProductController extends Controller
         $model_action = "delete";
         $model_id = $product->id;
 
-
-
         $this->notification->send($type, $by ,$data, $model_action, $model_id);
-        //
+
         $product->delete();
 
         Session::flash('success', 'Este producto se eliminó exitosamente.');
@@ -540,9 +711,8 @@ class ProductController extends Controller
         return view('wecommerce::back.products.index')->with('products', $products);
     }
 
-      public function filter($order , $filter)
+    public function filter($order , $filter)
     {
-
         $products = Product::orderBy($filter, $order)->paginate(15);
 
         if ($filter == 'sku' && $order == 'desc') {
